Commit 1.4

+++ src/main/java/com/iablonski/springboot/shop/spring_online_shop/domain/OrderDetails.java
@@ -27,7 +27,7 @@ public class OrderDetails {
     @ManyToOne
     @JoinColumn(name = "product_id")
     private Product product;
-    private BigDecimal amount;
-    private BigDecimal price;
+    private double amount;
+    private double price;
 
 }
